SYNCHRONIZED ROBOT AND PIRATE MOVEMENT!!!!!!!!!!
